Reorganize code into subfolders
