+++ Sources/JsonModel/Deprecated/IdentifiableInterfaceSerializer.swift
