+++ Sources/JsonModel/Json/JsonValue.swift
@@ -4,16 +4,6 @@
 
 import Foundation
 
-/// Protocol for converting an object to a dictionary representation. This is included for
-/// reverse-compatiblility to older implementations that are not Swift `Codable` and instead
-/// use a dictionary representation. Additionally, this can be used to implement Kotlin-Native
-/// serializable objects that do not conform to the Codable protocol.
-public protocol DictionaryRepresentable {
-    
-    /// Return the dictionary representation for this object.
-    func jsonDictionary() throws -> [String : JsonSerializable]
-}
-
 /// Protocol for converting objects to JSON serializable objects.
 public protocol JsonValue {
