+++ Sources/JsonModel/JsonSchema/JsonSchema.swift
