+++ Sources/JsonModel/PolymorphicSerialization/ResourceInfo.swift
